finish cap04 Funcionalidades da aplicação
