+++ src/Pages/Home/Components/CountDown/index.tsx
@@ -1,0 +1,14 @@
+import { CountdownContainer, Separator } from "./styles";
+
+export function CountDown() {
+  return (
+    <CountdownContainer>
+      <span>{minutes[0]}</span>
+      <span>{minutes[1]}</span>
+      <Separator>:</Separator>
+      <span>{seconds[0]}</span>
+      <span>{seconds[1]}</span>
+    </CountdownContainer>
+  );
+}
+
