+++ src/Pages/Home/Components/NewCycleForm/index.tsx
@@ -1,0 +1,37 @@
+import { FormContainer, TaskInput, MinutesAmountInput } from "./styles";
+
+export function NewCycleForm() {
+  return (
+    <FormContainer>
+      <label htmlFor="task">Vou trabalhar em</label>
+      <TaskInput
+        list="taskSugestions"
+        placeholder="Dê um nome para o seu projeto"
+        disabled={!!activeCycle}
+        {...register("task")}
+      />
+      <datalist id="taskSugestions">
+        <option value="01" />
+        <option value="A" />
+        <option value="2" />
+        <option value="B" />
+        <option value="José" />
+      </datalist>
+
+      <label htmlFor="minutesAmount">durante</label>
+      <MinutesAmountInput
+        id="minutesAmount"
+        type="number"
+        placeholder="00"
+        step={5}
+        min={5}
+        max={60}
+        disabled={!!activeCycle}
+        {...register("minutesAmount", { valueAsNumber: true })}
+      />
+
+      <span>minutos.</span>
+    </FormContainer>
+  );
+}
+
