+++ src/Pages/Home/Home.tsx
@@ -1,18 +1,17 @@
-import { Play } from "phosphor-react";
+import { HandPalm, Play } from "phosphor-react";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import * as zod from "zod";
 
 import {
   HomeContainer,
-  FormContainer,
-  CountdownContainer,
-  Separator,
   StartCountDownButton,
-  MinutesAmountInput,
-  TaskInput,
+  StopCountDownButton,
 } from "./styles";
-import { useState } from "react";
+import { useEffect, useState } from "react";
+import { differenceInSeconds } from "date-fns";
+import { NewCycleForm } from "./Components/NewCycleForm";
+import { CountDown } from "./Components/CountDown";
 
 const newCycleFormValidationSchema = zod.object({
   task: zod.string().min(1, "Informe a tarefa"),
@@ -26,6 +25,9 @@ interface Cycle {
   id: string;
   task: string;
   minutesAmount: number;
+  startDate: Date;
+  interrupedDate?: Date;
+  finishedDate?: Date;
 }
 
 type NewCycleFormData = zod.infer<typeof newCycleFormValidationSchema>;
@@ -43,6 +45,43 @@ export function Home() {
     },
   });
 
+  const activeCycle = cycles.find((cycle) => cycle.id === activeCycleId);
+
+  const totalSeconds = activeCycle ? activeCycle.minutesAmount * 60 : 0;
+
+  useEffect(() => {
+    let interval: number;
+
+    if (activeCycle) {
+      interval = setInterval(() => {
+        const secondsDifefrence = differenceInSeconds(
+          new Date(),
+          activeCycle.startDate
+        );
+
+        if (secondsDifefrence >= totalSeconds) {
+          setCycles((state) =>
+            state.map((cycle) => {
+              if (cycle.id === activeCycleId) {
+                return { ...cycle, finishedDate: new Date() };
+              } else {
+                return cycle;
+              }
+            })
+          );
+
+          setamountSecondsPassed(totalSeconds);
+          clearInterval(interval);
+        } else {
+          setamountSecondsPassed(secondsDifefrence);
+        }
+      }, 1000);
+    }
+    return () => {
+      clearInterval(interval);
+    };
+  }, [activeCycle, totalSeconds, activeCycleId]);
+
   function handleCreateNewCycle(data: NewCycleFormData) {
     const id = String(new Date().getTime());
 
@@ -50,17 +89,30 @@ export function Home() {
       id,
       task: data.task,
       minutesAmount: data.minutesAmount,
+      startDate: new Date(),
     };
 
     setCycles((state) => [...cycles, newCycle]);
     setActiveCycleId(id);
+    setamountSecondsPassed(0);
 
     reset();
   }
 
-  const activeCycle = cycles.find((cycle) => cycle.id === activeCycleId);
+  function handleInterrupedCycle() {
+    setCycles((state) =>
+      state.map((cycle) => {
+        if (cycle.id === activeCycleId) {
+          return { ...cycle, interrupedDate: new Date() };
+        } else {
+          return cycle;
+        }
+      })
+    );
 
-  const totalSeconds = activeCycle ? activeCycle.minutesAmount * 60 : 0;
+    setActiveCycleId(null);
+  }
+
   const currentSeconds = activeCycle ? totalSeconds - amountSecondsPassed : 0;
 
   const minutesAmount = Math.floor(currentSeconds / 60);
@@ -72,49 +124,28 @@ export function Home() {
   const task = watch("task");
   const isSubmitDisabled = !task;
 
+  useEffect(() => {
+    if (activeCycle) {
+      document.title = `${minutes}: ${seconds}`;
+    }
+  }, [minutes, seconds, activeCycle]);
+
   return (
     <HomeContainer>
       <form onSubmit={handleSubmit(handleCreateNewCycle)} action="">
-        <FormContainer>
-          <label htmlFor="task">Vou trabalhar em</label>
-          <TaskInput
-            list="taskSugestions"
-            placeholder="Dê um nome para o seu projeto"
-            {...register("task")}
-          />
-          <datalist id="taskSugestions">
-            <option value="01" />
-            <option value="A" />
-            <option value="2" />
-            <option value="B" />
-            <option value="José" />
-          </datalist>
-
-          <label htmlFor="minutesAmount">durante</label>
-          <MinutesAmountInput
-            id="minutesAmount"
-            type="number"
-            placeholder="00"
-            step={5}
-            min={5}
-            max={60}
-            {...register("minutesAmount", { valueAsNumber: true })}
-          />
-
-          <span>minutos.</span>
-        </FormContainer>
-
-        <CountdownContainer>
-          <span>{minutes[0]}</span>
-          <span>{minutes[1]}</span>
-          <Separator>:</Separator>
-          <span>{seconds[0]}</span>
-          <span>{seconds[1]}</span>
-        </CountdownContainer>
-
-        <StartCountDownButton disabled={isSubmitDisabled} type="submit">
-          <Play />
-        </StartCountDownButton>
+        <NewCycleForm />
+        <CountDown />
+        {activeCycle ? (
+          <StopCountDownButton type="button" onClick={handleInterrupedCycle}>
+            <HandPalm size={24} />
+            Interromper
+          </StopCountDownButton>
+        ) : (
+          <StartCountDownButton disabled={isSubmitDisabled} type="submit">
+            <Play />
+            Começar
+          </StartCountDownButton>
+        )}
       </form>
     </HomeContainer>
   );
