+++ src/Pages/Home/styles.ts
@@ -15,77 +15,7 @@ export const HomeContainer = styled.main`
     }
 `
 
-export const FormContainer = styled.div`
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: 0.5rem;
-    color: ${(props) => props.theme['gray-100']};
-    font-size: 1.125rem;
-    font-weight: bold;
-    flex-wrap: wrap;
-`
-
-const BaseInput = styled.input`
-    background: transparent;
-    height: 2.5rem;
-    border: 0;
-    border-bottom: 2px solid ${(props) => props.theme['gray-500']};
-    font-weight: bold;
-    font-size: 1.125rem;
-    padding:  0 0.5rem;
-    color: ${(props) => props.theme['gray-100']};
-
-    &::placeholder {
-      color: ${(props) => props.theme['gray-500']};
-    }
-
-    &:focus {
-        box-shadow: none;
-        border-color: ${(props) => props.theme['green-500']};
-    }
-`;
-
-export const TaskInput = styled(BaseInput)`
-    flex: 1;
-
-    &::-webkit-calendar-picker-indicator {
-        display: none !important;
-    }
-`;
-
-export const MinutesAmountInput = styled(BaseInput)`
-    width: 4rem;
-`;
-
-export const CountdownContainer = styled.div`
-    font-family: "Roboto Mono", monospace;
-    font-size: 10rem;
-    line-height: 8rem;
-    color: ${(props) => props.theme['gray-100']};
-
-    display: flex;
-    gap: 1rem;
-
-    span {
-        background: ${(props) => props.theme['gray-700']};
-        padding: 2rem 1rem;
-        border-radius: 8px;
-    }
-`
-
-export const Separator = styled.div`
-    padding: 2rem 0;
-    color: ${(props) => props.theme['green-500']};
-
-    width: 4rem;
-    overflow: hidden;
-    display: flex;
-    justify-content: center;
-`
-
-export const StartCountDownButton = styled.button`
+export const BaseCountDownButton = styled.button`
     width: 100%;
     border: 0;
     padding: 1rem;
@@ -94,21 +24,32 @@ export const StartCountDownButton = styled.button`
     display: flex;
     align-items: center;
     justify-content: center;
+    
 
     gap: 0.5rem;
     font-weight: bold;
+    color: ${(props) => props.theme['gray-100']};
 
     cursor: pointer;
-
-    background: ${(props) => props.theme['green-500']};
-    color: ${(props) => props.theme['gray-100']};
 
     &:disabled {
         opacity: 0.7;
         cursor: not-allowed;
     }
+`
+
+export const StartCountDownButton = styled(BaseCountDownButton)`
+    background: ${(props) => props.theme['green-500']};
 
     &:not(:disabled):hover {
         background: ${(props) => props.theme['green-700']};
+    }
+`
+
+export const StopCountDownButton = styled(BaseCountDownButton)`
+    background: ${(props) => props.theme['red-500']};
+
+    &:not(:disabled):hover {
+        background: ${(props) => props.theme['red-700']};
     }
 `
